Database design
My database contains 2 tables that have both a one-to-many and many-to-one relationship because a single user can make many events, and many events can belong to one user. The connection between the events and the user table is shown in the events table with that last column called “user_id”. This is the foreign key to the users table’s primary key.